+++ models/staging/sap_adventure_works/stg_sap__productcosthistory.sql
@@ -4,7 +4,7 @@ with
             cast(productid as int) as productid
             , cast(startdate as timestamp) as startdate
             , cast(enddate as timestamp) as enddate
-            , cast(standardcost as float) as standardcost
+            , cast(standardcost as decimal) as standardcost
             , cast(modifieddate as timestamp) as modifieddate
         from {{ source('sap_adventure_works', 'productcosthistory') }}
     )
