+++ models/staging/sap_adventure_works/stg_sap__salesorderdetail.sql
@@ -7,8 +7,8 @@ with
             , cast(orderqty as int) as orderqty
             , cast(productid as int) as productid
             , cast(specialofferid as int) as specialofferid
-            , cast(unitprice as float) as unitprice
-            , cast(unitpricediscount as float) as unitpricediscount
+            , cast(unitprice as decimal) as unitprice
+            , cast(unitpricediscount as decimal) as unitpricediscount
             , cast(rowguid as string) as rowguid
             , cast(modifieddate as timestamp) as modifieddate
         from {{ source('sap_adventure_works', 'salesorderdetail') }}
