+++ models/staging/sap_adventure_works/stg_sap__shipmethod.sql
@@ -3,8 +3,8 @@ with
         select
             cast(shipmethodid as int) as shipmethodid
             , cast(name as string) as name
-            , cast(shipbase as float) as shipbase
-            , cast(shiprate as float) as shiprate
+            , cast(shipbase as decimal) as shipbase
+            , cast(shiprate as decimal) as shiprate
             , cast(rowguid as string) as rowguid
             , cast(modifieddate as timestamp) as modifieddate
         from {{ source('sap_adventure_works', 'shipmethod') }}
